Further refactored components for the devtools panel UI. Added github logo and link in the footer.
