+++ src/devpanel/components/App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import './App.css';
 
 import Header from './Header/Header.js';
-import Main from './Main.js';
+import Main from './Main/Main.js';
 import Footer from './Footer/Footer.js'
 
 class App extends Component {
