+++ src/devpanel/components/MezzuriteInfo/MezzuriteDetails.css
@@ -1,0 +1,4 @@
+#mezzurite-package span, 
+#mezzurite-version span {
+    font-weight: bold;
+}
