+++ src/devpanel/components/MezzuriteInfo/MezzuriteDetails.js
@@ -1,11 +1,16 @@
 import React from 'react';
+import './MezzuriteDetails.css'
+import Container from '../Container.js';
 
 const MezzuriteDetails = (props) => (
-    <div className="MezzuriteInfo">
-        <div id="mezzurite-found">{props.mezzuriteFoundMessage}</div>
-        <div id="mezzurite-package">{props.mezzuritePackage}</div>
-        <div id="mezzurite-version">{props.mezzuriteVersion}</div>
-    </div>
+    <Container title="Mezzurite Framework Details">
+        <div id="mezzurite-package">
+            <span>Mezzurite Package Name: </span>{props.mezzuritePackage}
+        </div>
+        <div id="mezzurite-version">
+            <span>Mezzurite Package Version: </span>{props.mezzuriteVersion}
+        </div>
+    </Container>
 );
 
 export default MezzuriteDetails;
