+++ src/devpanel/components/MezzuriteInfo/MezzuriteInfo.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
-import Container from './Container.js';
-import MezzuriteInspector from '../services/MezzuriteInspector.js';
+import MezzuriteDetails from './MezzuriteDetails.js';
+import MezzuriteTimings from './MezzuriteTimings.js';
+import MezzuriteInspector from '../../services/MezzuriteInspector.js';
 
 class MezzuriteInfo extends Component {
     constructor(props) {
@@ -45,24 +46,15 @@ class MezzuriteInfo extends Component {
     }
 
     render() {
-        const timingItems = this.state.timings.map(timing => 
-            <li>{timing.metricType + ", " + timing.value + ", " + timing.data}</li>
-        );
         return (
             <div className="MezzuriteInfo">
-                <Container title="Mezzurite Framework Details">
-                    <div id="mezzurite-package">
-                        <span>Mezzurite Package Name: </span>{this.state.mezzuritePackage}
-                    </div>
-                    <div id="mezzurite-version">
-                        <span>Mezzurite Package Version: </span>{this.state.mezzuriteVersion}
-                    </div>
-                </Container>
-                <Container title="Mezzurite Timings">
-                    <ul id="timings">
-                        {timingItems}
-                    </ul>
-                </Container>
+                <MezzuriteDetails
+                    mezzuritePackage={this.state.mezzuritePackage}
+                    mezzuriteVersion={this.state.mezzuriteVersion}
+                />
+                <MezzuriteTimings 
+                    timings={this.state.timings}
+                />
             </div>
         );
     }
