+++ src/devpanel/index.js
@@ -7,7 +7,6 @@ import './index.css';
 
 import App from './components/App.js';
 
-
 // The only reason for using React Router DOM is to have ALT for the chrome extension
 // MemoryRouter is used because this is not a traditional browser environment
 ReactDOM.render((
@@ -16,22 +15,3 @@ ReactDOM.render((
     // </MemoryRouter>
     ),document.getElementById('root')
 );
-
-////////////////////////////////
-// Chrome stuff while I figure out how to refactor it nicely...
-////////////////////////////////
-
-// import { getMezzuriteObjectAsync } from './agent.js'
-
-// getMezzuriteObjectAsync()
-//     .then(value => {
-//         if (value === undefined) {
-//             // Mezzurite not found, at least for now
-//             // TODO do a 10-second timeout counter decrement for controlling the number of calls.
-
-//         } else {
-
-//         }
-
-
-//     });
